+++ test/table_test.dart
@@ -1,4 +1,6 @@
 import 'package:build_html/src/table.dart';
+import 'package:build_html/src/table_cell.dart';
+import 'package:build_html/src/table_row.dart';
 import 'package:test/test.dart';
 
 void tests() {
@@ -11,7 +13,10 @@ void tests() {
     });
     test('test caption', () {
       var table = Table()..withCaption('A demo table');
-      expect(table.toHtml(), '<table><thead/><tbody/><caption>A demo table</caption></table>');
+      expect(
+        table.toHtml(),
+        '<table><thead/><tbody/><caption>A demo table</caption></table>',
+      );
     });
     test('test header row', () {
       var table = Table()..withHeaderRow(['a', 'b']);
@@ -22,6 +27,18 @@ void tests() {
     });
     test('test data row', () {
       var table = Table()..withBodyRow(['a', 'b']);
+      expect(
+        table.toHtml(),
+        '<table><thead/><tbody><tr><td>a</td><td>b</td></tr></tbody></table>',
+      );
+    });
+    test('test custom data row', () {
+      var table = Table();
+      table.addCustomBodyRow(
+        TableRow()
+          ..withCell(TableCell(TableCellType.data)..withRaw('a'))
+          ..withCell(TableCell(TableCellType.data)..withRaw('b')),
+      );
       expect(
         table.toHtml(),
         '<table><thead/><tbody><tr><td>a</td><td>b</td></tr></tbody></table>',